+++ macro/cfa.py
@@ -321,4 +321,4 @@ class CatacombsFrostAwakened(Dungeon):
       util.dice_dungeon()
       util.log_action(util.MSG_END_DG)
       util.log_time()
-      util.wait(3)
+      util.wait(1)