+++ macro/ci.py
@@ -274,4 +274,4 @@ class ChaosInfinity(Dungeon):
       util.dice_dungeon()
       util.log_action(util.MSG_END_DG)
       util.log_time()
-      util.wait(3)
+      util.wait(1)