+++ macro/hva.py
@@ -701,4 +701,4 @@ class HazardousValleyAwakened(Dungeon):
       util.dice_dungeon()
       util.log_action(util.MSG_END_DG)
       util.log_time()
-      util.wait(3)
+      util.wait(1)