+++ macro/hvenh.py
@@ -847,4 +847,4 @@ class HazardousValley(Dungeon):
       util.dice_dungeon()
       util.log_action(util.MSG_END_DG)
       util.log_time()
-      util.wait(3)
+      util.wait(1)