+++ macro/hw.py
@@ -1563,4 +1563,4 @@ class HolyWindmill(Dungeon):
       util.dice_dungeon()
       util.log_action(util.MSG_END_DG)
       util.log_time()
-      util.wait(3)
+      util.wait(1)