+++ macro/lha.py
@@ -438,4 +438,4 @@ class LavaHellfireAwakened(Dungeon):
       util.dice_dungeon()
       util.log_action(util.MSG_END_DG)
       util.log_time()
-      util.wait(3)
+      util.wait(1)