+++ macro/pca.py
@@ -630,4 +630,4 @@ class PanicCaveAwakened(Dungeon):
       util.dice_dungeon()
       util.log_action(util.MSG_END_DG)
       util.log_time()
-      util.wait(3)
+      util.wait(1)