+++ macro/s1p.py
@@ -1183,4 +1183,4 @@ class SienaB1FPrideus(Dungeon):
       util.force_exit_dungeon()
       util.log_action(util.MSG_END_DG)
       util.log_time()
-      util.wait(3)
+      util.wait(1)