+++ macro/sca.py
@@ -416,4 +416,4 @@ class SteamerCrazyAwakened(Dungeon):
       util.dice_dungeon()
       util.log_action(util.MSG_END_DG)
       util.log_time()
-      util.wait(3)
+      util.wait(1)