+++ macro/tm.py
@@ -1171,4 +1171,4 @@ class TerminusMachina(Dungeon):
       util.dice_dungeon()
       util.log_action(util.MSG_END_DG)
       util.log_time()
-      util.wait(3)
+      util.wait(1)