+++ CHANGELOG.md
@@ -27,6 +27,7 @@ All notable changes to this project will be documented in this file.
 - [Application] Added back member status for party type automations (Beta Phase)
 - [Applicaiton] Added pet feature on beta phase, available now for testers
 - [Application] Added Chaos Infinity on beta phase, available now for testers
+- [Dungeon] Changed wait time before starting dungeon from 3 to 1
 
 ## v5.60 Fixes